# Summative-1-Kompella-Avineesh/Summative_1/src/main/java/com/company/Summative_1/models/Answer.java
package com.company.Summative_1.models;

public class Answer {

    private int id;
    private String question;
    private String answer;

    // Default constructor
    public Answer() {
    }

    public Answer(int id, String question, String answer) {
        this.id = id;
        this.question = question;
        this.answer = answer;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getQuestion() {
        return question;
    }

    public void setQuestion(String question) {
        this.question = question;
    }

    public String getAnswer() {
        return answer;
    }

    public void setAnswer(String answer) {
        this.answer = answer;
    }
}